finnish-chatbot-cloud/requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
anyio==4.8.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
decorator==5.2.1
fastapi>=0.109.2
ffmpeg-python==0.2.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2024.12.0
future==1.0.0
gTTS==2.5.4
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.6
playaudio==1.0.5
pooch==1.8.2
psutil==6.1.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn>=0.27.1
wheel==0.45.1
whisper==1.1.10
whispercpp==0.0.17

